Split oversized runtime viewer files
