+++ .pm/tasks/task_dbd38392feb44872ab7a2f5c2c0e6b25.execution.md
@@ -528,3 +528,15 @@ Example:
   - `./scripts/cargo-dev.sh test -p oasis7 runtime_gameplay_action_claim_uses_bound_player_agent_as_actor`: passed; 1 passed, 1160 filtered out in `src/lib.rs`, remaining test binaries filtered to 0.
   - `env -u RUSTC_WRAPPER cargo fmt --check && git diff --check`: passed.
   - `npm --prefix crates/oasis7_viewer run build:software-safe:bundle && node crates/oasis7_viewer/scripts/finalize-software-safe-build.mjs`: passed; finalized `crates/oasis7_viewer/viewer.js`.
+
+## 2026-06-29 18:03:18 CST / tpm
+- 完成内容: Fixed PR #740 `required-gate` failure from `check-rust-file-size.sh`.
+- CI Blocker:
+  - GitHub required-gate failed because current oversized scan was not empty: `crates/oasis7/src/viewer/runtime_live/control_plane.rs` had 1207 lines and `crates/oasis7/src/viewer/runtime_live/tests/snapshot_progress.rs` had 1391 lines.
+- Patch:
+  - Moved agent chat intent parsing into `crates/oasis7/src/viewer/runtime_live/agent_chat_intent.rs`, reducing `control_plane.rs` below the 1200-line gate.
+  - Moved empty-world / first-Agent claim snapshot tests into `crates/oasis7/src/viewer/runtime_live/tests/snapshot_progress_empty_world.rs`, reducing `snapshot_progress.rs` below the 1200-line gate.
+- Fresh Verification:
+  - `env -u RUSTC_WRAPPER cargo fmt --check && ./scripts/check-rust-file-size.sh && git diff --check`: passed; oversized code files=0, test files=0.
+  - `./scripts/cargo-dev.sh test -p oasis7 existing_world --lib`: passed; 3 passed, 1158 filtered out.
+  - `./scripts/cargo-dev.sh test -p oasis7 runtime_gameplay_action_claim_uses_bound_player_agent_as_actor`: passed; 1 passed, 1160 filtered out in `src/lib.rs`, remaining test binaries filtered to 0.
