+++ crates/oasis7/src/viewer/runtime_live/tests.rs
@@ -21,6 +21,7 @@ mod industrial_progression;
 mod prompt_control;
 mod provider_settings;
 mod snapshot_progress;
+mod snapshot_progress_empty_world;
 #[path = "tests_support.rs"]
 mod tests_support;
